Backend/requirements.txt
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.7
colorama==0.4.6
elevenlabs==2.44.0
filelock==3.29.0
fsspec==2026.3.0
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.12
Jinja2==3.1.6
llvmlite==0.47.0
MarkupSafe==3.0.3
more-itertools==11.0.2
mpmath==1.3.0
networkx==3.6.1
numba==0.65.0
numpy==2.4.4
ollama==0.6.1
openai-whisper==20250625
psycopg2==2.9.12
pydantic==2.13.3
pydantic_core==2.46.3
python-dotenv==1.2.2
regex==2026.4.4
requests==2.33.1
setuptools==81.0.0
SQLAlchemy==2.0.49
sympy==1.14.0
tiktoken==0.12.0
torch==2.11.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
